# storage: allow to Seek from some point and backwards