… pkg/core/native_neo_test.go
… pkg/core/storage/storeandbatch_test.go