Close #2187.

TODO:

- [x] Allow to seek backwards from some point (do we need it?).
- [x] Use `SeekFrom` and `SeekBackwards` where appropriate.
- [x] Benchmarks.
- [x] Fix commets.
- [x] Optimize time of `getGasPerVote` (something is wrong with it, execution time after optimisation is larger than before).