@roman-khimov 2021-12-29
IIRC, backwards seek is useful for NEO voter GAS calculation.