@AnnaShaleva 2021-12-29
> IIRC, backwards seek is useful for NEO voter GAS calculation.

Isn't it implemented in ee67c8b474a060cb26f7f97d22a22e402d4fc187?

The first TODO is mostly about the case when we need to seek backwards *and* starting from some point, i.e. *non-strict backwards seeking*. I wasn't able to find where we can make use of it.