@AnnaShaleva 2022-01-10
As a result, seeking starting from some point, forwards/backwards and their combinations are implemented, but some tests need to be added and investigation of `getGasPerVote` benchmark behaviour need to be done.