@AnnaShaleva 2022-01-11
Finally, here are the benchmark results. I'll attach them to the commit message after the review.
```
benchstat BenchmarkNEO_GetGASPerVote_master_all_records BenchmarkNEO_GetGASPerVote_feature_all_records 
name                                              old time/op    new time/op    delta
NEO_GetGASPerVote/MemPS_10_reward_records-8         27.2µs ± 3%    29.3µs ± 3%   +7.92%  (p=0.000 n=10+9)
NEO_GetGASPerVote/MemPS_100_reward_records-8         192µs ± 3%     221µs ± 9%  +14.93%  (p=0.000 n=9+10)
NEO_GetGASPerVote/MemPS_1000_reward_records-8       6.40ms ±48%    7.89ms ±93%     ~     (p=0.436 n=10+10)
NEO_GetGASPerVote/MemPS_10000_reward_records-8      34.2ms ±12%    33.4ms ± 6%     ~     (p=0.971 n=10+10)
NEO_GetGASPerVote/BoltPS_10_reward_records-8        25.8µs ± 4%    26.3µs ± 2%   +2.01%  (p=0.041 n=10+7)
NEO_GetGASPerVote/BoltPS_100_reward_records-8       94.2µs ±31%    63.4µs ±15%  -32.70%  (p=0.000 n=10+10)
NEO_GetGASPerVote/BoltPS_1000_reward_records-8       417µs ±18%     253µs ±13%  -39.34%  (p=0.000 n=10+10)
NEO_GetGASPerVote/BoltPS_10000_reward_records-8     3.24ms ± 3%    1.77ms ±12%  -45.17%  (p=0.000 n=9+10)
NEO_GetGASPerVote/LevelPS_10_reward_records-8       26.9µs ±10%    29.5µs ± 9%   +9.94%  (p=0.003 n=9+10)
NEO_GetGASPerVote/LevelPS_100_reward_records-8      94.0µs ±10%    90.7µs ±16%     ~     (p=0.280 n=10+10)
NEO_GetGASPerVote/LevelPS_1000_reward_records-8      562µs ± 8%     453µs ± 6%  -19.43%  (p=0.000 n=9+10)
NEO_GetGASPerVote/LevelPS_10000_reward_records-8    4.83ms ±10%    3.34ms ±12%  -30.69%  (p=0.000 n=10+10)

name                                              old alloc/op   new alloc/op   delta
NEO_GetGASPerVote/MemPS_10_reward_records-8         5.63kB ± 0%    5.26kB ± 0%   -6.50%  (p=0.000 n=10+10)
NEO_GetGASPerVote/MemPS_100_reward_records-8        37.4kB ± 0%    29.7kB ± 0%  -20.54%  (p=0.000 n=9+10)
NEO_GetGASPerVote/MemPS_1000_reward_records-8        371kB ± 2%     280kB ± 8%  -24.51%  (p=0.000 n=7+10)
NEO_GetGASPerVote/MemPS_10000_reward_records-8      5.02MB ± 1%    4.21MB ± 1%  -16.21%  (p=0.000 n=10+10)
NEO_GetGASPerVote/BoltPS_10_reward_records-8        5.94kB ± 1%    5.53kB ± 1%   -6.80%  (p=0.000 n=10+8)
NEO_GetGASPerVote/BoltPS_100_reward_records-8       30.5kB ±13%    19.9kB ± 6%  -34.74%  (p=0.000 n=10+10)
NEO_GetGASPerVote/BoltPS_1000_reward_records-8       219kB ± 3%     131kB ± 5%  -40.50%  (p=0.000 n=10+10)
NEO_GetGASPerVote/BoltPS_10000_reward_records-8     2.32MB ± 8%    1.40MB ± 7%  -39.87%  (p=0.000 n=10+10)
NEO_GetGASPerVote/LevelPS_10_reward_records-8       6.38kB ± 0%    6.28kB ± 2%   -1.59%  (p=0.007 n=9+10)
NEO_GetGASPerVote/LevelPS_100_reward_records-8      28.6kB ± 3%    22.8kB ± 9%  -20.14%  (p=0.000 n=10+10)
NEO_GetGASPerVote/LevelPS_1000_reward_records-8      210kB ± 2%     149kB ± 2%  -29.08%  (p=0.000 n=10+10)
NEO_GetGASPerVote/LevelPS_10000_reward_records-8    2.07MB ± 4%    1.40MB ± 3%  -32.50%  (p=0.000 n=10+10)

name                                              old allocs/op  new allocs/op  delta
NEO_GetGASPerVote/MemPS_10_reward_records-8            136 ± 1%       104 ± 1%  -23.69%  (p=0.000 n=10+10)
NEO_GetGASPerVote/MemPS_100_reward_records-8           839 ± 2%       454 ± 3%  -45.87%  (p=0.000 n=9+10)
NEO_GetGASPerVote/MemPS_1000_reward_records-8        9.00k ± 1%     5.03k ± 1%  -44.07%  (p=0.000 n=10+10)
NEO_GetGASPerVote/MemPS_10000_reward_records-8       90.6k ± 0%     50.6k ± 0%  -44.15%  (p=0.000 n=10+9)
NEO_GetGASPerVote/BoltPS_10_reward_records-8           142 ± 0%       110 ± 1%  -22.96%  (p=0.000 n=10+10)
NEO_GetGASPerVote/BoltPS_100_reward_records-8          879 ± 0%       484 ± 0%  -44.87%  (p=0.000 n=10+10)
NEO_GetGASPerVote/BoltPS_1000_reward_records-8       8.18k ± 0%     4.15k ± 1%  -49.30%  (p=0.000 n=10+10)
NEO_GetGASPerVote/BoltPS_10000_reward_records-8      83.3k ± 2%     42.9k ± 2%  -48.52%  (p=0.000 n=10+10)
NEO_GetGASPerVote/LevelPS_10_reward_records-8          150 ± 1%       118 ± 3%  -21.09%  (p=0.000 n=10+10)
NEO_GetGASPerVote/LevelPS_100_reward_records-8         862 ± 1%       488 ± 1%  -43.35%  (p=0.000 n=10+10)
NEO_GetGASPerVote/LevelPS_1000_reward_records-8      8.13k ± 0%     4.21k ± 0%  -48.18%  (p=0.000 n=10+9)
NEO_GetGASPerVote/LevelPS_10000_reward_records-8     81.5k ± 2%     41.5k ± 1%  -49.02%  (p=0.000 n=10+10)
```