@AnnaShaleva 2022-01-11
And here are the neo-bench results for four and single nodes for LevelDB and BoltDB with standard configuration. Memory consumption and TPS for LevelDB is better with the feature branch, but for BoltDB the results are controversial. I'm thinking about running neo-bench with larger block count for BoltDB.
![cpu_4_nodes_10_wrk](https://user-images.githubusercontent.com/22775709/148983519-581a9f94-e7fe-4412-aca9-90e46e62d55a.png)
![cpu_single_10_wrk](https://user-images.githubusercontent.com/22775709/148983524-d1698856-d045-4114-9c49-a547c1acc664.png)
![mem_4_nodes_10_wrk](https://user-images.githubusercontent.com/22775709/148983526-b04b1069-243c-4d8a-8c4b-bafb90f561c3.png)
![mem_single_10_wrk](https://user-images.githubusercontent.com/22775709/148983529-07acb0e9-8f6e-4fbe-b0ea-8fa5e486bf25.png)
![ms_per_block_4_nodes_10_wrk](https://user-images.githubusercontent.com/22775709/148983533-91348ace-0f01-4e31-b426-6d367f893f1a.png)
![ms_per_block_single_10_wrk](https://user-images.githubusercontent.com/22775709/148983535-446b4f02-ded6-4fd4-affe-a135e4e74b29.png)
![tpb_4_nodes_10_wrk](https://user-images.githubusercontent.com/22775709/148983537-04250b21-dd23-45c7-949c-277ed895b983.png)
![tpb_single_10_wrk](https://user-images.githubusercontent.com/22775709/148983539-109a8939-e75b-4275-be79-bb7f5ff73c70.png)
![tps_4_nodes_10_wrk](https://user-images.githubusercontent.com/22775709/148983540-c94acd4f-58cd-4a68-bb89-5dbde8849a6c.png)
![tps_single_10_wrk](https://user-images.githubusercontent.com/22775709/148983542-3b4b60a9-4841-4379-bce4-85752f2b82f6.png)

[Go4x1_wrk_10_feature.log](https://github.com/nspcc-dev/neo-go/files/7848322/Go4x1_wrk_10_feature.log)
[Go4x1_wrk_10_feature_bolt.log](https://github.com/nspcc-dev/neo-go/files/7848323/Go4x1_wrk_10_feature_bolt.log)
[Go4x1_wrk_10_master.log](https://github.com/nspcc-dev/neo-go/files/7848324/Go4x1_wrk_10_master.log)
[Go4x1_wrk_10_master_bolt.log](https://github.com/nspcc-dev/neo-go/files/7848325/Go4x1_wrk_10_master_bolt.log)
[GoSingle_wrk_10_feature.log](https://github.com/nspcc-dev/neo-go/files/7848327/GoSingle_wrk_10_feature.log)
[GoSingle_wrk_10_feature_bolt.log](https://github.com/nspcc-dev/neo-go/files/7848328/GoSingle_wrk_10_feature_bolt.log)
[GoSingle_wrk_10_master.log](https://github.com/nspcc-dev/neo-go/files/7848329/GoSingle_wrk_10_master.log)
[GoSingle_wrk_10_master_bolt.log](https://github.com/nspcc-dev/neo-go/files/7848330/GoSingle_wrk_10_master_bolt.log)