@roman-khimov 2022-01-11
> neo-bench results for four and single nodes for LevelDB and BoltDB with standard configuration

Single non-voting account? How about a 1000 of voting accounts?